Backend/requirements.txt
google-cloud-datastore>=1.7.3
gunicorn>=19.9.0
PyJWT>=1.7.1
python-dateutil>=2.8.0
pytz>=2019.1
requests>=2.22.0
twilio>=6.27.1
ics>=0.4
sendgrid>=6.0.5

-e git://github.com/savvybit/ndb-orm.git@develop#egg=ndb_orm
